feat: 🎸 async plugin init
